moved bitmap code to separate class
