+++ CMakeLists.txt
@@ -10,7 +10,7 @@ file(GLOB HQX_SOURCES
 	"source/HQ2x.cc"
 	"source/HQ3x.cc")
 
-add_library(hqx ${HQX_SOURCES})
+add_library(hqx ${HQX_SOURCES} "include/Bitmap.h")
 set_target_properties(hqx PROPERTIES
     OUTPUT_NAME "hqx"
     ARCHIVE_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}"
@@ -18,7 +18,7 @@ set_target_properties(hqx PROPERTIES
     RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}"
     PREFIX "lib" )
 
-add_executable(test "source/main.cc")
+add_executable(test "source/main.cc" "include/Bitmap.h")
 target_link_libraries(test hqx)
 set_target_properties(test PROPERTIES
     OUTPUT_NAME "test"
